requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug>=3.0.1
python-dotenv==1.0.1
google-genai==0.3.0
